+++ plone/app/contentrules/tests/test_action_mail.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from email import message_from_string
-from email.Message import Message
 from zope.component import getUtility, getMultiAdapter, getSiteManager
 from zope.component.interfaces import IObjectEvent
 from zope.interface import implements
@@ -12,7 +11,9 @@ from plone.contentrules.engine.interfaces import IRuleStorage
 from plone.contentrules.rule.interfaces import IRuleAction, IExecutable
 from plone.registry.interfaces import IRegistry
 
+from Acquisition import aq_base
 from Products.CMFPlone.interfaces.controlpanel import IMailSchema
+from Products.CMFPlone.tests.utils import MockMailHost
 from Products.MailHost.interfaces import IMailHost
 from Products.MailHost.MailHost import MailHost
 
@@ -24,18 +25,6 @@ class DummyEvent(object):
 
     def __init__(self, object):
         self.object = object
-
-
-class DummyMailHost(MailHost):
-    meta_type = 'Dummy Mail Host'
-
-    def __init__(self, id):
-        self.id = id
-        self.sent = []
-
-    def _send(self, mfrom, mto, messageText, *args, **kw):
-        msg = message_from_string(messageText)
-        self.sent.append(msg)
 
 
 class TestMailAction(ContentRulesTestCase):
@@ -56,6 +45,22 @@ class TestMailAction(ContentRulesTestCase):
             self.portal.portal_membership.addMember(id, 'secret', roles, [])
             member = self.portal.portal_membership.getMemberById(id)
             member.setMemberProperties({'fullname': fname, 'email': email})
+
+    def _setup_mockmail(self):
+        sm = getSiteManager(self.portal)
+        sm.unregisterUtility(provided=IMailHost)
+        dummyMailHost = MockMailHost('MailHost')
+        sm.registerUtility(dummyMailHost, IMailHost)
+        self.portal._original_MailHost = self.portal.MailHost
+        self.portal.MailHost = dummyMailHost
+        return dummyMailHost
+
+    def _teardown_mockmail(self):
+        self.portal.MailHost = self.portal._original_MailHost
+        sm = getSiteManager(context=self.portal)
+        sm.unregisterUtility(provided=IMailHost)
+        sm.registerUtility(aq_base(self.portal._original_MailHost),
+                           provided=IMailHost)
 
     def testRegistered(self):
         element = getUtility(IRuleAction, name='plone.actions.Mail')
@@ -96,23 +101,21 @@ class TestMailAction(ContentRulesTestCase):
         self.assertTrue(isinstance(editview, MailEditFormView))
 
     def testExecute(self):
-        self.loginAsPortalOwner()
+        self.loginAsPortalOwner()   # this avoids sending mail as currentuser@foobar.com
         self.portal.portal_membership.getAuthenticatedMember().setProperties(
             email='currentuser@foobar.com')
-        sm = getSiteManager(self.portal)
-        sm.unregisterUtility(provided=IMailHost)
-        dummyMailHost = DummyMailHost('dMailhost')
-        sm.registerUtility(dummyMailHost, IMailHost)
+        dummyMailHost = self._setup_mockmail()
         e = MailAction()
         e.source = "$user_email"
-        e.recipients = "bar@foo.be, bar@foo.be, $reviewer_emails, $manager_emails, $member_emails"
+        e.recipients = "bar@foo.be, bar@foo.be, $reviewer_emails, $manager_emails, $member_emails"   # noqa
         e.message = "P\xc3\xa4ge '${title}' created in ${url} !".decode('utf-8')
         ex = getMultiAdapter((self.folder, e, DummyEvent(self.folder.d1)),
                              IExecutable)
         ex()
-        self.assertTrue(isinstance(dummyMailHost.sent[0], Message))
-
-        sent_mails = dict([(mailSent.get('To'), mailSent) for mailSent in dummyMailHost.sent])
+        sent_mails = {}
+        for mail_sent in dummyMailHost.messages:
+            mail_sent_msg = message_from_string(mail_sent)
+            sent_mails[mail_sent_msg.get('To')] = mail_sent_msg
 
         mailSent = sent_mails['bar@foo.be']
         self.assertEqual('text/plain; charset="utf-8"',
@@ -133,13 +136,10 @@ class TestMailAction(ContentRulesTestCase):
         self.assertEqual(
             set(["bar@foo.be", "user@one.com", "user@two.com", "user@three.com", "user@four.com", ]),  # noqa
             set(sent_mails.keys()))
+        self._teardown_mockmail()
 
     def testExecuteNoSource(self):
-        self.loginAsPortalOwner()
-        sm = getSiteManager(self.portal)
-        sm.unregisterUtility(provided=IMailHost)
-        dummyMailHost = DummyMailHost('dMailhost')
-        sm.registerUtility(dummyMailHost, IMailHost)
+        dummyMailHost = self._setup_mockmail()
         e = MailAction()
         e.recipients = 'bar@foo.be,foo@bar.be'
         e.message = 'Document created !'
@@ -155,8 +155,7 @@ class TestMailAction(ContentRulesTestCase):
         mail_settings.email_from_address = 'manager@portal.be'
         mail_settings.email_from_name = u'plone@rulez'
         ex()
-        self.assertTrue(isinstance(dummyMailHost.sent[0], Message))
-        mailSent = dummyMailHost.sent[0]
+        mailSent = message_from_string(dummyMailHost.messages[0])
         self.assertEqual('text/plain; charset="utf-8"',
                          mailSent.get('Content-Type'))
         self.assertEqual("bar@foo.be", mailSent.get('To'))
@@ -164,13 +163,10 @@ class TestMailAction(ContentRulesTestCase):
                          mailSent.get('From'))
         self.assertEqual("Document created !",
                          mailSent.get_payload(decode=True))
+        self._teardown_mockmail()
 
     def testExecuteMultiRecipients(self):
-        self.loginAsPortalOwner()
-        sm = getSiteManager(self.portal)
-        sm.unregisterUtility(provided=IMailHost)
-        dummyMailHost = DummyMailHost('dMailhost')
-        sm.registerUtility(dummyMailHost, IMailHost)
+        dummyMailHost = self._setup_mockmail()
         e = MailAction()
         e.source = 'foo@bar.be'
         e.recipients = 'bar@foo.be,foo@bar.be'
@@ -178,27 +174,23 @@ class TestMailAction(ContentRulesTestCase):
         ex = getMultiAdapter((self.folder, e, DummyEvent(self.folder.d1)),
                              IExecutable)
         ex()
-        self.assertEqual(len(dummyMailHost.sent), 2)
-        self.assertTrue(isinstance(dummyMailHost.sent[0], Message))
-        mailSent = dummyMailHost.sent[0]
+        self.assertEqual(len(dummyMailHost.messages), 2)
+        mailSent = message_from_string(dummyMailHost.messages[0])
         self.assertEqual('text/plain; charset="utf-8"',
                          mailSent.get('Content-Type'))
         self.assertEqual('bar@foo.be', mailSent.get('To'))
         self.assertEqual('foo@bar.be', mailSent.get('From'))
         self.assertEqual('Document created !', mailSent.get_payload(decode=True))
-        mailSent = dummyMailHost.sent[1]
+        mailSent = message_from_string(dummyMailHost.messages[1])
         self.assertEqual('text/plain; charset="utf-8"',
                          mailSent.get('Content-Type'))
         self.assertEqual('foo@bar.be', mailSent.get('To'))
         self.assertEqual('foo@bar.be', mailSent.get('From'))
         self.assertEqual('Document created !', mailSent.get_payload(decode=True))
+        self._teardown_mockmail()
 
     def testExecuteExcludeActor(self):
-        self.loginAsPortalOwner()
-        sm = getSiteManager(self.portal)
-        sm.unregisterUtility(provided=IMailHost)
-        dummyMailHost = DummyMailHost('dMailhost')
-        sm.registerUtility(dummyMailHost, IMailHost)
+        dummyMailHost = self._setup_mockmail()
         self.portal.portal_membership.getAuthenticatedMember().setProperties(
             email='currentuser@foobar.com')
         e = MailAction()
@@ -209,18 +201,15 @@ class TestMailAction(ContentRulesTestCase):
         ex = getMultiAdapter((self.folder, e, DummyEvent(self.folder.d1)),
                              IExecutable)
         ex()
-        self.assertEqual(len(dummyMailHost.sent), 1)
+        self.assertEqual(len(dummyMailHost.messages), 1)
 
-        mailSent = dummyMailHost.sent[0]
+        mailSent = message_from_string(dummyMailHost.messages[0])
         self.assertEqual("bar@foo.be", mailSent.get('To'))
+        self._teardown_mockmail()
 
     def testExecuteNoRecipients(self):
         # no recipient
-        self.loginAsPortalOwner()
-        sm = getSiteManager(self.portal)
-        sm.unregisterUtility(provided=IMailHost)
-        dummyMailHost = DummyMailHost('dMailhost')
-        sm.registerUtility(dummyMailHost, IMailHost)
+        dummyMailHost = self._setup_mockmail()
         e = MailAction()
         e.source = 'foo@bar.be'
         e.recipients = ''
@@ -228,7 +217,8 @@ class TestMailAction(ContentRulesTestCase):
         ex = getMultiAdapter((self.folder, e, DummyEvent(self.folder.d1)),
                              IExecutable)
         ex()
-        self.assertEqual(len(dummyMailHost.sent), 0)
+        self.assertEqual(len(dummyMailHost.messages), 0)
+        self._teardown_mockmail()
 
     @unittest.skip('Monkey patching does not work well with mocking. Needs fixing.')
     def testExecuteBadMailHost(self):